Implement reading binary blobs from file
It would be useful to be able to save a binary blob in raw format to a
file and run the parsing and processing on this saved data later. E.g.
for debugging, development or help requests by others. It would also
allow to save you unparsable data now and try parsing with a future
version of the program - e.g. after a bug was fixed.